+++ packages/babel-plugin/src/class-names/__tests__/tagged-template-expression.test.ts
@@ -15,7 +15,7 @@ describe('ClassNames used with a css tagged template expression', () => {
 
       expect(actual).toIncludeMultiple([
         'const _ = "._1wybgktf{font-size:20px}"',
-        'className={"_1wybgktf"}',
+        'className={ax(["_1wybgktf"])}',
       ]);
     });