+++ packages/css/src/plugins/atomicify-rules.ts
@@ -3,6 +3,7 @@ import type { Plugin, ChildNode, Declaration, Container, Rule, AtRule } from 'po
 import { rule } from 'postcss';
 
 interface PluginOpts {
+  classNameCompressionMap?: { [index: string]: string };
   callback?: (className: string) => void;
 }
 
@@ -76,20 +77,28 @@ const replaceNestingSelector = (selector: string, parentClassName: string) => {
  * @param node
  */
 const buildAtomicSelector = (node: Declaration, opts: AtomicifyOpts) => {
+  const { classNameCompressionMap } = opts;
   const selectors: string[] = [];
 
   (opts.selectors || ['']).forEach((selector) => {
     const normalizedSelector = normalizeSelector(selector);
-    const className = atomicClassName(node, {
+    const fullClassName = atomicClassName(node, {
       ...opts,
       selectors: [normalizedSelector],
     });
-    const replacedSelector = replaceNestingSelector(normalizedSelector, className);
 
-    selectors.push(replacedSelector);
+    const compressedClassName =
+      classNameCompressionMap && classNameCompressionMap[fullClassName.slice(1)];
+
+    if (compressedClassName) {
+      // Use compressed class name if compressedClassName is available
+      selectors.push(replaceNestingSelector(normalizedSelector, compressedClassName));
+    } else {
+      selectors.push(replaceNestingSelector(normalizedSelector, fullClassName));
+    }
 
     if (opts.callback) {
-      opts.callback(className);
+      opts.callback(fullClassName);
     }
   });