+++ packages/parcel-transformer/src/types.ts
@@ -39,4 +39,12 @@ export interface ParcelTransformerOpts extends BabelPluginOpts {
    * Default to `false`
    */
   addComponentName?: boolean;
+
+  /**
+   * A map holds the key-value pairs between full Atomic class names and the compressed ones
+   * i.e. { '_aaaabbbb': 'a' }
+   *
+   * Default to `undefined`
+   */
+  classNameCompressionMap?: { [index: string]: string };
 }