+++ packages/react/src/runtime/__tests__/ax.test.ts
@@ -1,70 +1,81 @@
 import ax from '../ax';
 
 describe('ax', () => {
-  it('should join single classes together', () => {
-    const result = ax(['foo', 'bar']);
+  const isEnabled: boolean = (() => false)();
 
-    expect(result).toEqual('foo bar');
-  });
-
-  it('should join multi classes together', () => {
-    const result = ax(['foo baz', 'bar']);
-
-    expect(result).toEqual('foo baz bar');
-  });
-
-  it('should remove undefined', () => {
-    const result = ax(['foo', 'bar', undefined]);
-
-    expect(result).toEqual('foo bar');
-  });
-
-  it('should ensure the last atomic declaration of a single group wins', () => {
-    const result = ax(['_aaaabbbb', '_aaaacccc']);
-
-    expect(result).toEqual('_aaaacccc');
-  });
-
-  it('should ensure the last atomic declaration of many single groups wins', () => {
-    const result = ax(['_aaaabbbb', '_aaaacccc', '_aaaadddd', '_aaaaeeee']);
-
-    expect(result).toEqual('_aaaaeeee');
-  });
-
-  it('should ensure the last atomic declaration of a multi group wins', () => {
-    const result = ax(['_aaaabbbb _aaaacccc']);
-
-    expect(result).toEqual('_aaaacccc');
-  });
-
-  it('should ensure the last atomic declaration of many multi groups wins', () => {
-    const result = ax(['_aaaabbbb _aaaacccc _aaaadddd _aaaaeeee']);
-
-    expect(result).toEqual('_aaaaeeee');
-  });
-
-  it('should not remove any atomic declarations if there are no duplicate groups', () => {
-    const result = ax(['_aaaabbbb', '_bbbbcccc']);
-
-    expect(result).toEqual('_aaaabbbb _bbbbcccc');
-  });
-
-  it('should not apply conditional class', () => {
-    const isEnabled: boolean = (() => false)();
-    const result = ax([isEnabled && 'foo', 'bar']);
-
-    expect(result).toEqual('bar');
-  });
-
-  it('should ignore non atomic declarations', () => {
-    const result = ax(['hello_there', 'hello_world']);
-
-    expect(result).toEqual('hello_there hello_world');
-  });
-
-  it('should ignore non atomic declarations when atomic declarations exist', () => {
-    const result = ax(['hello_there', 'hello_world', '_aaaabbbb']);
-
-    expect(result).toEqual('hello_there hello_world _aaaabbbb');
+  it.each([
+    ['should handle empty array', [], undefined],
+    ['should handle array with undefined', [undefined], undefined],
+    ['should join single classes together', ['foo', 'bar'], 'foo bar'],
+    ['should join multi classes together', ['foo baz', 'bar'], 'foo baz bar'],
+    ['should remove undefined', ['foo', 'bar', undefined], 'foo bar'],
+    [
+      'should ensure the last atomic declaration of a single group wins',
+      ['_aaaabbbb', '_aaaacccc'],
+      '_aaaacccc',
+    ],
+    [
+      'should ensure the last atomic declaration of a single group with short class name wins',
+      ['_aaaabbbb', '_aaaacccc', '_aaaa_a'],
+      'a',
+    ],
+    [
+      'should ensure the last atomic declaration of many single groups wins',
+      ['_aaaabbbb', '_aaaacccc', '_aaaadddd', '_aaaaeeee'],
+      '_aaaaeeee',
+    ],
+    [
+      'should ensure the last atomic declaration of many single groups with short class name wins',
+      ['_aaaabbbb', '_aaaacccc', '_aaaa_a', '_aaaa_b'],
+      'b',
+    ],
+    [
+      'should ensure the last atomic declaration of a multi group wins',
+      ['_aaaabbbb _aaaacccc'],
+      '_aaaacccc',
+    ],
+    [
+      'should ensure the last atomic declaration of a multi group with short class name wins',
+      ['_aaaa_e', '_aaaabbbb _aaaacccc'],
+      '_aaaacccc',
+    ],
+    [
+      'should ensure the last atomic declaration of many multi groups wins',
+      ['_aaaabbbb _aaaacccc _aaaadddd _aaaaeeee'],
+      '_aaaaeeee',
+    ],
+    [
+      'should ensure the last atomic declaration of many multi groups with short class name wins',
+      ['_aaaabbbb', '_aaaa_a', '_bbbb_b', '_ddddcccc'],
+      'a b _ddddcccc',
+    ],
+    [
+      'should not remove any atomic declarations if there are no duplicate groups',
+      ['_aaaabbbb', '_bbbbcccc'],
+      '_aaaabbbb _bbbbcccc',
+    ],
+    [
+      'should not remove any atomic declarations if there are short class name and no duplicate groups',
+      ['_eeee_e', '_aaaabbbb', '_bbbbcccc'],
+      'e _aaaabbbb _bbbbcccc',
+    ],
+    ['should not apply conditional class', [isEnabled && 'foo', 'bar'], 'bar'],
+    [
+      'should ignore non atomic declarations',
+      ['hello_there', 'hello_world'],
+      'hello_there hello_world',
+    ],
+    [
+      'should ignore non atomic declarations when atomic declarations exist',
+      ['hello_there', 'hello_world', '_aaaabbbb'],
+      'hello_there hello_world _aaaabbbb',
+    ],
+    [
+      'should ignore non atomic declarations when atomic declarations with short class name exist',
+      ['hello_there', 'hello_world', '_aaaa_a'],
+      'hello_there hello_world a',
+    ],
+  ])('%s', (_, params, result) => {
+    expect(result).toEqual(ax(params));
   });
 });