+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @compiled/react
 
+## 0.12.0
+
+### Minor Changes
+
+- a41e41e6: Update monorepo node version to v18, and drop support for node v12
+- f9c957ef: Add an option to compress class names based on "classNameCompressionMap", which is provided by library consumers.
+  Add a script to generate compressed class names.
+
 ## 0.11.4
 
 ### Patch Changes